+++ simpeg_drivers-assets/uijson/induced_polarization_2d_forward.ui.json
@@ -19,12 +19,17 @@
     },
     "line_object": {
         "association": "Cell",
-        "dataType": "Referenced",
+        "dataType": [
+            "Integer",
+            "Referenced"
+        ],
         "group": "Survey",
         "main": true,
         "label": "Line ID",
         "parent": "data_object",
         "value": "",
+        "optional": true,
+        "enabled": false,
         "tooltip": "Selects the data representing the different lines in the survey."
     },
     "line_id": {
@@ -35,6 +40,8 @@
         "value": 1,
         "optional": true,
         "enabled": false,
+        "dependency": "line_object",
+        "dependencyType": "enabled",
         "tooltip": "Selects the line of data to be processed."
     },
     "chargeability_channel_bool": true,
