+++ simpeg_drivers/utils/nested.py
@@ -64,10 +64,10 @@ def create_nested_mesh(
     if base_mesh.dim == 1:
         return base_mesh
 
-    return create_nested_2dmesh(survey, base_mesh)
+    return create_nested_2d_tensor(survey, base_mesh)
 
 
-def create_nested_2dmesh(
+def create_nested_2d_tensor(
     survey: BaseSurvey,
     base_mesh: TensorMesh,
 ) -> TensorMesh:
@@ -275,7 +275,7 @@ def create_simulation(
 
     if isinstance(simulation, BaseEM1DSimulation):
         local_mesh = simulation.layers_mesh
-        actives = np.ones(simulation.layers_mesh.n_cells, dtype=bool)
+        local_actives = np.ones(simulation.layers_mesh.n_cells, dtype=bool)
         model_slice = np.arange(
             indices, simulation.mesh.n_cells, simulation.mesh.shape_cells[0]
         )[::-1]
@@ -303,26 +303,28 @@ def create_simulation(
                     3 if getattr(simulation, "model_type", None) == "vector" else 1
                 ),
             )
-            actives = mapping.local_active
+            local_actives = mapping.local_active
         # For DCIP-2D
         else:
-            actives_2d = simulation.active_cells.reshape(
-                simulation.mesh.shape_cells, order="F"
-            )
-            local_active = actives_2d[simulation.mesh.parts == tile_id, :]
-            actives = local_active.flatten(order="F")
-
             # Create a projection from the global active cells to the local active cells
+            active_mesh_part = np.isin(
+                simulation.mesh.parts, np.unique(local_survey.line_ids)
+            )
             n_actives = simulation.active_cells.sum()
             activate_ind = np.zeros(simulation.mesh.n_cells, dtype=int)
             activate_ind[np.where(simulation.active_cells)[0]] = np.arange(n_actives)
             activate_ind = activate_ind.reshape(simulation.mesh.shape_cells, order="F")
-            local_active_ind = activate_ind[
-                simulation.mesh.parts == tile_id, :
-            ].flatten(order="F")[actives]
-            mapping = maps.Projection(n_actives, local_active_ind)
 
-    n_actives = int(actives.sum())
+            actives_2d = simulation.active_cells.reshape(
+                simulation.mesh.shape_cells, order="F"
+            )
+            local_actives = actives_2d[active_mesh_part, :].flatten(order="F")
+            local_active_inds = activate_ind[active_mesh_part, :].flatten(order="F")[
+                local_actives
+            ]
+            mapping = maps.Projection(n_actives, local_active_inds)
+
+    n_actives = int(local_actives.sum())
     if getattr(simulation, "_chiMap", None) is not None:
         if simulation.model_type == "vector":
             kwargs["chiMap"] = maps.IdentityMap(nP=n_actives * 3)
@@ -330,22 +332,24 @@ def create_simulation(
         else:
             kwargs["chiMap"] = maps.IdentityMap(nP=n_actives)
 
-        kwargs["active_cells"] = actives
+        kwargs["active_cells"] = local_actives
 
     if getattr(simulation, "_rhoMap", None) is not None:
         kwargs["rhoMap"] = maps.IdentityMap(nP=n_actives)
-        kwargs["active_cells"] = actives
+        kwargs["active_cells"] = local_actives
 
     if getattr(simulation, "_sigmaMap", None) is not None:
         kwargs["sigmaMap"] = maps.ExpMap(local_mesh) * maps.InjectActiveCells(
-            local_mesh, actives, value_inactive=np.log(1e-8)
+            local_mesh, local_actives, value_inactive=np.log(1e-8)
         )
 
     if getattr(simulation, "_etaMap", None) is not None:
-        kwargs["etaMap"] = maps.InjectActiveCells(local_mesh, actives, value_inactive=0)
+        kwargs["etaMap"] = maps.InjectActiveCells(
+            local_mesh, local_actives, value_inactive=0
+        )
         proj = maps.InjectActiveCells(
             local_mesh,
-            actives,
+            local_actives,
             value_inactive=1e-8,
         )
         kwargs["sigma"] = proj * mapping * simulation.sigma[simulation.active_cells]
@@ -439,6 +443,10 @@ def create_survey(
 
     slice_inds = slice_from_ordering(survey, indices, channel=channel)
     new_survey.ordering = survey.ordering[slice_inds, :]
+
+    if hasattr(survey, "line_ids"):
+        new_survey.line_ids = survey.line_ids[slice_inds]
+
     if hasattr(survey, "dobs") and survey.dobs is not None:
         # Return the subset of data that belongs to the tile
